Cleanup significantly improves behavior.  Aligned variable type definitions for timers from byte and double to unsigned int.  Refactored from arduino Delay() to delayMicroseconds() (since Delay is blocking and uses timer resources and delayMicroseconds is not - that lets this library play much nicer with other code).  Also noticed & addressed cases of interrupts being disabled but never re-enabled.

+++ PS2X_lib.h
@@ -226,8 +226,8 @@ volatile uint32_t *		_att_lport_clr;
 uint16_t 				_dat_mask; 
 volatile uint32_t *		_dat_lport;
 #endif
-unsigned long last_read;
-byte read_delay;
+unsigned long last_read;    // milliseconds
+unsigned long read_delay;   // microseconds
 byte controller_type;
 boolean en_Rumble;
 boolean en_Pressures;
